requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
environs==8.0.0
glqiwiapi==1.1.3
psycopg2-binary==2.9.1
requests==2.26.0
Telethon==1.23.0

